signing identity with connection button
