--- ceramic.config.json
@@ -1,3 +1,0 @@
-{
-  "did:3:kjzl6cwe1jw146dnkczbmmca1p0wkiyiiark51440c1ui3o6vvy0img4wqi816o": "k2t6wyfsu4pfz1b3gyk3p0as1jd3ionoxf7c5tnwxud55lt4bazizj5zofguyk"
-}
