+++ pages/_app.js
@@ -4,15 +4,14 @@ import { store } from '../app/store'
 import { Provider } from 'react-redux'
 import { Provider as CeramicProvider } from '@self.id/react'
 import XmtpProvider from '../xmtp/components/XmtpProvider.tsx'
-// import WalletProvider from '../xmtp/components/WalletProvider.tsx'
 
 function MyApp({ Component, pageProps }) {
   return (
     <Provider store={store}>
       <CeramicProvider client={{ ceramic: 'testnet-clay' }}>
-          <XmtpProvider>
+        <XmtpProvider>
             <Component {...pageProps} />
-          </XmtpProvider>
+        </XmtpProvider>
       </CeramicProvider>
     </Provider>
   )
