+++ pages/rotarydial/index.js
@@ -5,9 +5,12 @@ import ProfileCard from '../../components/Profile/ProfileCard'
 import { useViewerRecord } from '@self.id/react'
 import { useDispatch, useSelector } from 'react-redux'
 import { setContacts, setSyncedCeramic } from '../../app/contactSlice'
-import { useEffect } from 'react'
+import { useCallback, useEffect, useRef } from 'react'
 import { dummyProfile } from '../../components/Profile/ProfileCard'
 import ProfileModal from '../../components/ProfileModal'
+import {WalletProvider} from '../../xmtp/components/WalletProvider.tsx'
+
+
 
 const Profile = () => {
   // This is the entrypoint to the users database.
@@ -17,6 +20,8 @@ const Profile = () => {
   const store = useSelector((state) => state.contact)
   const evmStore = useSelector((state) => state.evm)
   const dispatch = useDispatch()
+
+
 
   useEffect(() => {
     if (!store.hasInitialRecord && evmStore.connected && record.content) {
@@ -44,6 +49,9 @@ const Profile = () => {
     store.isSyncedCeramic,
   ])
 
+  
+  
+
   // async function clickContacts() {
   //   record.set({
   //     'contacts': {
@@ -55,6 +63,8 @@ const Profile = () => {
   // }
 
   return (
+    <WalletProvider>
+
     <div className="bg-mybg-light dark:bg-mybg-dark min-h-screen flex flex-col ">
       <Head />
       {/* <button className="bg-mybg-light dark:bg-mybg-dark min-h-screen flex flex-col " onClick={clickContacts} >
@@ -63,6 +73,8 @@ const Profile = () => {
       <Sidebar />
       <ProfileModal />
     </div>
+    </WalletProvider>
+
   )
 }
